Rename go-tests to gotests
