+++ enterprise-suite/gotests/resources/app_service_with_only_endpoints.yaml
