+++ enterprise-suite/gotests/resources/app_with_service.yaml
