+++ enterprise-suite/gotests/util/util.go
